criado uma VIEW chamada faturamento_atual que exibe os faturamentos atuais, conforme especificado no requisito 6

+++ desafio1.sql
@@ -6,7 +6,7 @@ USE SpotifyClone ;
 CREATE TABLE IF NOT EXISTS SpotifyClone.Plano (
     plano_id INT NOT NULL AUTO_INCREMENT,
     plano VARCHAR(45) NOT NULL,
-    valor_plano INT NOT NULL,
+    valor_plano DOUBLE NOT NULL,
     PRIMARY KEY (plano_id)
 );
